+++ work-shop/src/containers/create.jsx
@@ -1,7 +1,8 @@
 import React, { Component } from 'react'
-import { TitleC, InputT1, InputP1, InputN, InputE, BtNext} from '../style/StyleGlobal';
+import { TitleC, InputT1, InputP1, InputN, InputE, BtNext, TexBT} from '../style/style';
 
 import axios from 'axios'
+import { InputP2 } from './../style/style';
 
 const baseUrl = 'https://app-liveedu.herokuapp.com/Usuarios';
 class create extends Component {
@@ -12,6 +13,7 @@ class create extends Component {
          email:'',
          telefono:'',
          password:'',
+         foto_perfil:'',
          id:''
       }
    }
@@ -26,9 +28,22 @@ class create extends Component {
          email:this.state.form.email,
          telefono:this.state.form.telefono,
          password: this.state.form.password,
-         
-         coins:"0"
-      }
+         foto_perfil: this.state.form.foto_perfil,
+         seguidores: "10.7M",
+         likes: "108.3M",
+         post: {
+            poster: "https://res.cloudinary.com/academiageek/image/upload/v1659129680/workshop/Group_11_rvmco0.png",
+            decription: "  Lorem ipsum dolor sit amet, consectetur adipiscing elit. Facilisi ullamcorper aliquam augue fermentum, vel, risus, adipiscing.Tellus vestibulum magnis ut elit, vitae mattis sapien. Laoreet nam",
+            Me_enecanta: "300K",
+            comentarios: "87K",
+            mensajes: "10K"
+         },
+         fotos: {
+            foto1: "https://res.cloudinary.com/academiageek/image/upload/v1659129679/workshop/Group_15_u4mh7t.png",
+            foto2: "https://res.cloudinary.com/academiageek/image/upload/v1659129679/workshop/Group_16_y794el.png",
+            foto3: "https://res.cloudinary.com/academiageek/image/upload/v1659129679/workshop/Group_17_i3giam.png",
+            foto4: "https://res.cloudinary.com/academiageek/image/upload/v1659129680/workshop/Group_19_kp7oqu.png"
+      }}
          ) 
         .catch(error =>{
          if (error.response) {
@@ -68,10 +83,11 @@ class create extends Component {
             <InputN  name="nombre" placeholder="nombre" size="2em" onChange={this.handleChange} />
             <InputE name="email" placeholder="email" size="2em"  onChange={this.handleChange} />
             <InputT1 name="telefono" placeholder="phone number" size="2em" onChange={this.handleChange}/>
-            <InputP1 name="password" placeholder="Password" size="2em" onChange={this.handleChange} />
+            <InputP1 name="password" type="password" placeholder="password" size="2em" onChange={this.handleChange} />
+            <InputP2 name="foto_perfil" type="img" placeholder="url imagen de perfil" size="2em" onChange={this.handleChange} />
          </form>
-
-         <BtNext onClick={() => this.registrer()}></BtNext> 
+         
+         <BtNext onClick={() => this.registrer()}><TexBT>crear</TexBT></BtNext> 
          
          </React.StrictMode>
       </div>
